requirements.txt
requests==2.31.0
jsonlines==4.0.0
pyyaml==6.0.1